Updated to use "subcommand" package.
We now have things behave in a more standard fashion, all arguments
follow the appropriate sub-command rather than being global.

The downside to this is that multiple commands (`serve` & `purge`)
need to have their DB specified to point to the same location.

That said this _is_ cleaner, so it stands.

This closes #9.