Store only RELATIVE filenames in our SQLite database.
Assuming we're launched like so:

        puppet-summary serve -prefix /srv/reports

We used to store reports in our SQLIte database such as:

        /srv/reports/ssh.example.com/12345
        /srv/reports/www.example.com/12345

Now we only store _relative_ paths, such as:

        www.example.com/1235
        ssh.example.com/1235

That means the user can change the prefix on the command-line,
`mv` the report-directory and everything "just works".  There is
a little more overhead in the lookup, but I think this is saner.