Graph failed/changed/unchanged over N days on index.
This changes the index-page from showing a donut of all the
states of the known-nodes _right now_ to showing the states
over the past few days (i.e. All data which has not been
pruned.)

This gives a visual history showing changes/trends, which is
neater.  We've also switched to using chart.js, because it is
less nosiy.

This closes #11, and updates #12.