Parse/store resources in a structure.
Rather than storing changed/failed/skipped resources in a mere
string we now store them in a structure:

* name
* file in the puppet-manifest which defined it
* line number in the file.

This allows the resource to be displayed more interestingly in the
report, and this closes #18.

At the same time I've reworked the header and the report-views to
be more "pretty".  Unrelated, logically, but since the point of
the structure was to make the display of the report more attractive
and useful it goes in at the same time.