Show relative-times in the reports.
This was a beast of a bug to deal with.  In short though I changed
the way that reports are stored:

* In the past we'd parse a report's time when it wa submitted.
  * We'd record that (parsed) time in the SQLite database.
* Now we don't do that.
  * Instead the report is assumed to have happened at the time the submission was received.

With that single difference we can calculate how long ago a run
happened solely on the report-server - since whatever time-zone it
is on is constant.

We've changed the reports to show the relative time, but if you hover
over the cell you'll see the absolute date&time.

This closes #28, and requires no changes to the database-schema,
no break, and no compatilibity issues.

**NOTE**: We're also now storing runs as reports/$host/$hash, since
we probably won't be married to the time in the future.