Correctly show relative time for orphaned-nodes.
My puppet-master broken, post-reboot, which left all my nodes
unable to talk to puppet-master ("Connection Refused").  Oops.

Checking on the state of my nodes I saw that all the hosts
were orphaned, as expected, but the relative time was empty
and unpopulated.

This commit ensures that the relative time is created even for
orphaned nodes.  e.g:

    attic.default.skx.uk0.bigv.io   orphaned        45 hours ago
    backup.vlan299.skx.uk0.bigv.io  orphaned        45 hours ago
    beta.vlan299.skx.uk0.bigv.io    orphaned        45 hours ago