Use javascript to linkify our table-rows.
This update changes the handling of links to using javascript
magic to make entire table-rows clickable - as we do in the other
views.

Making the change is __required__ because otherwise the "border"
we display atop the terms makes the links unclickable by normal
means.  (i.e. The coloured border, used to represent the %-total,
is drawn on top of the labels such as "unchanged", "failed", and
makes the links unclickable).

This closes #36