Don't cap node-reports to 50.
This closes #74, by removing the hardwired limit of node-reports
we display.

While it might be more flexible to allow a configurable limit
too many flags get annoying so I think we'll just remove it.

If there are too many reports for a user/node then I'd suggest
pruning more frequently.