Overhauled how static-resources are served
Rather than having a single handler for JS files, another
handler for CSS files, etc, we now have one unified handler
which is invoked for all static-resources.

We achieve this by adding a 404-handler, which first of all
checks whether the incoming-request can be satisfied via the
contents of our static-resources.  If so it will be served,
with an appropriate MIME-type, if not then a real 404
result will be returned to the caller.

The intention behind this change is that in the future we
can embed arbitrary (additional) static-resources without
the need to change our code.